Restructure test directory layout and rename test helpers
- Move test subject dirs (analysis, cli, formatters, mongo-shell,
  persistence, plugins) into test/tests/ to separate them from
  test infrastructure
- Rename test/utils/ → test/helpers/ to better describe its role
- Rename MongoShell.js → MongoShellRunner.js (clarifies it is an
  execution helper, not a test of the mongo-shell source module)
- Rename Tester.js → VarietyHarness.js and class Tester → VarietyHarness
- Rename JsonValidator.js → AnalysisResultsValidator.js and class
  JsonValidator → AnalysisResultsValidator
- Rename test-on-docker.sh → test-in-container.sh (supports Podman too)
- Rename npm script test:docker → test:container to match
- Update all import paths, JSDoc typedefs, and cross-references
  in test files, fixtures, package.json, ci.yml, and CONTRIBUTING.md
