+++ .github/workflows/ci.yml
@@ -95,7 +95,7 @@ jobs:
       - name: Expose GitHub Actions cache runtime
         uses: crazy-max/ghaction-github-runtime@04d248b84655b509d8c44dc1d6f990c879747487 # v4.0.0
       - run: mkdir -p docker/Dockerfiles
-      - run: npm run test:docker
+      - run: npm run test:container
         env:
           MONGODB_VERSION: ${{ matrix.mongodb-version }}
           NODEJS_VERSION: '22'
@@ -121,7 +121,7 @@ jobs:
       - name: Expose GitHub Actions cache runtime
         uses: crazy-max/ghaction-github-runtime@04d248b84655b509d8c44dc1d6f990c879747487 # v4.0.0
       - run: mkdir -p docker/Dockerfiles
-      - run: npm run test:docker
+      - run: npm run test:container
         env:
           MONGODB_VERSION: '8.0'
           NODEJS_VERSION: '24'
