+++ test/tests/analysis/PrototypeNamedKeysTest.js
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: © 2026 James Cropcho <numerate_penniless652@dralias.com>
-import Tester from '../utils/Tester.js';
+import VarietyHarness from '../../helpers/VarietyHarness.js';
 
 /** @typedef {import('mongodb').Document} MongoDocument */
 
-const test = new Tester('test', 'users');
+const test = new VarietyHarness('test', 'users');
 
 /** @type {MongoDocument[]} */
 const sampleData = [
