+++ test/tests/analysis/UnnamedObjectsAnalysisTest.js
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: © 2026 James Cropcho <numerate_penniless652@dralias.com>
-import Tester from '../utils/Tester.js';
-const test = new Tester('test', 'users');
+import VarietyHarness from '../../helpers/VarietyHarness.js';
+const test = new VarietyHarness('test', 'users');
 
 const sampleData = [
   {title:'Article 1', comments:[{author:'John', body:'it works', visible:true, '123key': '123value' }]},
