+++ test/tests/cli/ParametersParsingTest.js
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: © 2026 James Cropcho <numerate_penniless652@dralias.com>
 import assert from 'assert';
-import execute from '../utils/MongoShell.js';
-import Tester from '../utils/Tester.js';
-import sampleData from '../fixtures/SampleData.js';
+import execute from '../../helpers/MongoShellRunner.js';
+import VarietyHarness from '../../helpers/VarietyHarness.js';
+import sampleData from '../../fixtures/SampleData.js';
 
-const test = new Tester('test', 'users');
+const test = new VarietyHarness('test', 'users');
 const mongodbPort = Number(process.env['MONGODB_PORT'] || 27017);
 
 /**
