+++ test/tests/persistence/ContinuousLoggingTest.js
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: © 2026 James Cropcho <numerate_penniless652@dralias.com>
 import assert from 'assert';
-import Tester from '../utils/Tester.js';
-import sampleData from '../fixtures/SampleData.js';
+import VarietyHarness from '../../helpers/VarietyHarness.js';
+import sampleData from '../../fixtures/SampleData.js';
 
-const test = new Tester('test', 'users');
+const test = new VarietyHarness('test', 'users');
 
 const pattern = /^Found new key type "(.{1,})" type "(.{1,})"$/g;
 const expectedLines = [
