+++ test/tests/persistence/PersistenceTest.js
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: © 2026 James Cropcho <numerate_penniless652@dralias.com>
 import assert from 'assert';
-import Tester from '../utils/Tester.js';
-import sampleData from '../fixtures/SampleData.js';
+import VarietyHarness from '../../helpers/VarietyHarness.js';
+import sampleData from '../../fixtures/SampleData.js';
 
-/** @typedef {import('../utils/JsonValidator.js').VarietyResultRow} VarietyResultRow */
+/** @typedef {import('../../helpers/AnalysisResultsValidator.js').VarietyResultRow} VarietyResultRow */
 
-const test = new Tester('test', 'users');
+const test = new VarietyHarness('test', 'users');
 
 describe('Persistence of results', () => {
